netwerk/test/mochitests/test_signed_web_packaged_app_origin.html

<!DOCTYPE html>
<html>
<head>
  <title> Web packaged app </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="application/javascript;version=1.7">

var Cc = SpecialPowers.Cc;
var Ci = SpecialPowers.Ci;
var Cu = SpecialPowers.Cu;
var Cr = SpecialPowers.Cr;

SpecialPowers.pushPrefEnv(
  { "set": [["network.http.enable-packaged-apps", true],
            ["network.http.packaged-apps-developer-mode", true],,
            ["dom.mozBrowserFramesEnabled", true]] },
  () => SpecialPowers.pushPermissions([
    { "type": "browser", "allow": 1, "context": document }
  ], function() {
    runTest();
  }));

SimpleTest.waitForExplicitFinish();

function runTest() {
  var iframe = document.createElement("iframe");
  iframe.setAttribute('mozbrowser', 'true');
  iframe.setAttribute('remote', 'true');
  iframe.setAttribute("src", "http://example.org:80");

  iframe.addEventListener("mozbrowserloadend", function loadend(e) {
    iframe.removeEventListener("mozbrowserloadend", loadend);
    ok(true, "Got mozbrowserloadend");

    iframe.setAttribute("src", "http://mochi.test:8888/tests/netwerk/test/mochitests/signed_web_packaged_app.sjs!//scripts/app.js");

    iframe.addEventListener("mozbrowserloadend", function loadend(e) {
      iframe.removeEventListener("mozbrowserloadend", loadend);
        var mm = SpecialPowers.wrap(iframe)
                              .QueryInterface(Ci.nsIFrameLoaderOwner)
                              .frameLoader
                              .messageManager;

        var fsl = mm.QueryInterface(Ci.nsIFrameScriptLoader);

        mm.addMessageListener("my-e10s-extension-message", function(message) {
          let wrapper = message.objects.origin;
          ok(wrapper, "nodePrincipalOrigin:" + wrapper);
          SimpleTest.finish();
        });

        let frameScript =
`
function getNodePrincipalOrigin() {
  sendAsyncMessage("my-e10s-extension-message", {}, { origin: content.document.nodePrincipal.origin });
}
`;
        fsl.loadFrameScript("data:,(" + frameScript + ")()", true);
    });
  });

  document.body.appendChild(iframe);
}

</script>
</pre>
</body>
</html>
